Check if location is already displayed when displaying
the check is now done when asked for display info about a location and no more when trying to figure if location is on the map or not.
So that, when you ask for displaying places from a location, you don't have to thing about it, it is the responsabilité of the display to do that.