Remove coworking spaces.
I first imagined [devfriendlyplaces.net](http://devfriendlyplaces.net) as a list of places where I could go, seat myself at a table and start doing things and know if I will be able to plug my laptop, or order a tea, coffe, something to eat.
I wanted it to be a list of all those places not meant for work but where I could work while being surrounding by people living their life (or not).

Coworking spaces are not that kind of spaces. They are places meant for work. All people here are there to work. They surely are great places. But they are not for me.
I don't feel fine being asked what I'm doing, how long I will stay, why I'm leaving while I'm only here for 20min. They meant well, it has always been nice and friendly, and they sure are great places for some people. Just not for me, they are not what I'm looking for.

* Why now ?
I recently went to Paris with few spare hours to spend. I wanted to find a place to work on a pet project, and paris.devfriendlyplaces.net showed lot of points. That sounds great. All (but 2) were coworking spaces, places where I did not want to go, and so I suddenly feel lost and that some cleaning was required so devfriendlyplaces becomes what it meant to me, again. Here it is.